+++ src/main/java/com/example/domain/study/Schedule.java
@@ -1,17 +1,17 @@
-package com.example.domain;
+package com.example.domain.study;
 
 import com.example.domain.common.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.DynamicInsert;
@@ -24,20 +24,22 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class Vote extends BaseEntity {
-    @Id @GeneratedValue
+public class Schedule extends BaseEntity {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private Study study;
 
-    //private Member member;
-
     @Column(nullable = false)
     private String title;
 
     @Column(nullable = false)
-    private boolean isMultipleChoice;
+    private String location;
+
+    @Column(nullable = false)
+    private LocalDateTime staredAt;
 
     @Column(nullable = false)
     private LocalDateTime finishedAt;
