+++ src/main/java/com/example/domain/study/StudyLikedComment.java
@@ -1,15 +1,12 @@
-package com.example.domain;
+package com.example.domain.study;
 
 import com.example.domain.common.BaseEntity;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -25,28 +22,14 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class StudyPostComment extends BaseEntity {
+public class StudyLikedComment extends BaseEntity {
 
-    @Id @GeneratedValue
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    private StudyPost studyPost;
+    private StudyPostComment studyPostComment;
 
-    // @Column(nullable = false)
-    // private Member member;
-
-    @Column(nullable = false)
-    private String content;
-
-    @Column(nullable = false)
-    private Integer likeCount;
-
-    @OneToMany(mappedBy = "studyPostComment", cascade = CascadeType.ALL)
-    private List<StudyLikedComment> likedComments;
-
-
-
-
+    //private Member member;
 
 }
