+++ src/main/java/com/example/domain/study/StudyPostComment.java
@@ -1,8 +1,6 @@
-package com.example.domain;
+package com.example.domain.study;
 
 import com.example.domain.common.BaseEntity;
-import com.example.domain.enums.StudyState;
-import com.example.domain.enums.Theme;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -11,7 +9,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
-import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -28,27 +25,16 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class StudyPost extends BaseEntity {
+public class StudyPostComment extends BaseEntity {
 
     @Id @GeneratedValue
     private long id;
 
-    //private Member member;
-
     @ManyToOne(fetch = FetchType.LAZY)
-    private Study study;
+    private StudyPost studyPost;
 
-    @Column(nullable = false)
-    private boolean isAnnouncement;
-
-    @Column(nullable = false)
-    private LocalDateTime announcedAt;
-
-    @Column(nullable = false)
-    private Theme theme;
-
-    @Column(nullable = false)
-    private String title;
+    // @Column(nullable = false)
+    // private Member member;
 
     @Column(nullable = false)
     private String content;
@@ -56,16 +42,11 @@ public class StudyPost extends BaseEntity {
     @Column(nullable = false)
     private Integer likeCount;
 
-    @Column(nullable = false)
-    private Integer commentCount;
+    @OneToMany(mappedBy = "studyPostComment", cascade = CascadeType.ALL)
+    private List<StudyLikedComment> likedComments;
 
-    @OneToMany(mappedBy = "studyPost", cascade = CascadeType.ALL)
-    private List<StudyPostImage> images;
 
-    @OneToMany(mappedBy = "studyPost", cascade = CascadeType.ALL)
-    private List<StudyPostComment> comments;
 
-    @OneToMany(mappedBy = "studyPost", cascade = CascadeType.ALL)
-    private List<StudyLikedPost> likedPosts;
+
 
 }
