+++ src/main/java/com/example/domain/study/Vote.java
@@ -1,12 +1,13 @@
-package com.example.domain;
+package com.example.domain.study;
 
 import com.example.domain.common.BaseEntity;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,15 +23,22 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class StudyLikedPost extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+public class Vote extends BaseEntity {
+    @Id @GeneratedValue
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    private StudyPost studyPost;
+    private Study study;
 
     //private Member member;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private boolean isMultipleChoice;
+
+    @Column(nullable = false)
+    private LocalDateTime finishedAt;
 
 }
